app/requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
email-validator==1.1.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==1.10.1
Flask-Limiter==1.4
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
flask-talisman==0.8.1
Flask-WTF==1.2.1
greenlet==3.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.0.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
pexpect==4.9.0
ptyprocess==0.7.0
python-dotenv==0.19.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.25
Strip==0.0.1.dev14
stripe==7.0.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.0.1
wrapt==1.17.2
WTForms==2.3.3
gunicorn==20.1.0